moved run config stuff to folder
