+++ app/web_ui/src/lib/ui/run_config_component/prompt_type_selector.svelte
