+++ app/web_ui/src/routes/(app)/generate/[project_id]/[task_id]/+page.svelte
@@ -8,9 +8,9 @@
   import { page } from "$app/stores"
   import type { SampleDataNode } from "./gen_model"
   import GeneratedDataNode from "./generated_data_node.svelte"
-  import AvailableModelsDropdown from "$lib/ui/available_models_dropdown.svelte"
+  import AvailableModelsDropdown from "$lib/ui/run_config_component/available_models_dropdown.svelte"
   import { ui_state } from "$lib/stores"
-  import PromptTypeSelector from "$lib/ui/prompt_type_selector.svelte"
+  import PromptTypeSelector from "$lib/ui/run_config_component/prompt_type_selector.svelte"
   import FormContainer from "$lib/utils/form_container.svelte"
   import { type SampleData } from "./gen_model"
   import { indexedDBStore } from "$lib/stores/index_db_store"
